+++ x/crosschain/types/cctx_test.go
@@ -199,7 +199,6 @@ func TestCrossChainTx_AddOutbound(t *testing.T) {
 		require.Equal(t, cctx.GetCurrentOutboundParam().EffectiveGasPrice, sdkmath.NewInt(100))
 		require.Equal(t, cctx.GetCurrentOutboundParam().EffectiveGasLimit, uint64(20))
 		require.Equal(t, cctx.GetCurrentOutboundParam().ObservedExternalHeight, uint64(10))
-		require.Equal(t, cctx.CctxStatus.LastUpdateTimestamp, ctx.BlockHeader().Time.Unix())
 	})
 
 	t.Run("successfully get outbound tx for failed ballot without amount check", func(t *testing.T) {
@@ -220,7 +219,6 @@ func TestCrossChainTx_AddOutbound(t *testing.T) {
 		require.Equal(t, cctx.GetCurrentOutboundParam().EffectiveGasPrice, sdkmath.NewInt(100))
 		require.Equal(t, cctx.GetCurrentOutboundParam().EffectiveGasLimit, uint64(20))
 		require.Equal(t, cctx.GetCurrentOutboundParam().ObservedExternalHeight, uint64(10))
-		require.Equal(t, cctx.CctxStatus.LastUpdateTimestamp, ctx.BlockHeader().Time.Unix())
 	})
 
 	t.Run("failed to get outbound tx if amount does not match value received", func(t *testing.T) {