+++ zetaclient/chains/evm/signer/signer_admin_test.go
@@ -20,9 +20,8 @@ func TestSigner_SignAdminTx(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, nil)
 	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -89,10 +88,9 @@ func TestSigner_SignWhitelistERC20Cmd(t *testing.T) {
 	// Setup txData struct
 	cctx := getCCTX(t)
 
-	mockObserver, err := getNewEvmChainObserver(t, tss)
 	require.NoError(t, err)
 
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.NoError(t, err)
 	require.False(t, skip)
 
@@ -138,10 +136,9 @@ func TestSigner_SignMigrateERC20CustodyFundsCmd(t *testing.T) {
 	// Setup txData struct
 	cctx := getCCTX(t)
 
-	mockObserver, err := getNewEvmChainObserver(t, tss)
 	require.NoError(t, err)
 
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.NoError(t, err)
 	require.False(t, skip)
 
@@ -203,9 +200,8 @@ func TestSigner_SignMigrateTssFundsCmd(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
 	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)