+++ zetaclient/chains/evm/signer/signer_test.go
@@ -195,9 +195,7 @@ func TestSigner_SignOutbound(t *testing.T) {
 	// Setup txData struct
 
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
-	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, makeLogger(t))
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -237,7 +235,7 @@ func TestSigner_SignOutbound(t *testing.T) {
 		cctx.OutboundParams[0].GasPriorityFee = big.NewInt(priorityFee).String()
 
 		// Given outbound data
-		txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, makeLogger(t))
+		txData, skip, err := NewOutboundData(ctx, cctx, 123, makeLogger(t))
 		require.False(t, skip)
 		require.NoError(t, err)
 
@@ -270,9 +268,7 @@ func TestSigner_SignRevertTx(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
-	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -310,9 +306,7 @@ func TestSigner_SignCancelTx(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
-	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -350,9 +344,7 @@ func TestSigner_SignWithdrawTx(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
-	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -389,9 +381,7 @@ func TestSigner_SignERC20WithdrawTx(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, tss)
-	require.NoError(t, err)
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.False(t, skip)
 	require.NoError(t, err)
 
@@ -429,10 +419,7 @@ func TestSigner_BroadcastOutbound(t *testing.T) {
 
 	// Setup txData struct
 	cctx := getCCTX(t)
-	mockObserver, err := getNewEvmChainObserver(t, nil)
-	require.NoError(t, err)
-
-	txData, skip, err := NewOutboundData(ctx, cctx, mockObserver, 123, zerolog.Logger{})
+	txData, skip, err := NewOutboundData(ctx, cctx, 123, zerolog.Logger{})
 	require.NoError(t, err)
 	require.False(t, skip)